% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_seababel}
\alias{scale_color_seababel}
\title{ggplot2 color scale using SEA BABEL palette}
\usage{
scale_color_seababel(discrete = TRUE, ...)
}
\arguments{
\item{discrete}{Whether to use discrete or continuous scale.}

\item{...}{Additional arguments to pass to scale_color_manual or scale_color_gradientn}
}
\description{
ggplot2 color scale using SEA BABEL palette
}
