% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{sea_babel_palette}
\alias{sea_babel_palette}
\title{SeaBABEL Color Palette}
\usage{
sea_babel_palette(n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colors to return.}

\item{type}{Either "discrete" or "continuous".}
}
\value{
A vector of colors.
}
\description{
Custom color palette inspired by the Marine Bioacoustics and Behavioral Ecology Lab (SeaBABEL) at the University of New Hampshire.
}
\examples{
sea_babel_palette(5, type = "discrete")
sea_babel_palette(100, type = "continuous")
}
