% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seababel-package.R
\docType{package}
\name{seababel-package}
\alias{seababel}
\alias{seababel-package}
\title{seababel: A custom color palette inspired by the SeaBabel lab logo}
\description{
The \code{seababel} package provides a custom discrete and continuous color palette
inspired by the Marine Bioacoustics and Behavioral Ecology Lab (SeaBABEL)
at the University of New Hampshire.
}
\details{
It includes:
\itemize{
\item \code{sea_babel_palette()} to generate discrete or continuous palettes.
\item \code{scale_fill_seababel()} and \code{scale_color_seababel()} for use in ggplot2.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/faithfrings/seababel}
  \item Report bugs at \url{https://github.com/faithfrings/seababel/issues}
}

}
\author{
\strong{Maintainer}: Faith Frings \email{faith.frings@unh.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
